### 2020-05-21 #2191 @whedon
[PRE REVIEW]: SysIdentPy: A Python package for System Identification
**Submitting author:** @wilsonrljr (<a href="http://orcid.org/0000-0002-3263-1152">Wilson Rocha Lacerda Junior</a>)
**Repository:** <a href="https://github.com/wilsonrljr/sysidentpy" target ="_blank">https://github.com/wilsonrljr/sysidentpy</a>
**Version:** 0.1.1
