### 2020-05-28 #2191 @dpsanders
Hi @wilsonrljr, sorry for the delay.

I have a few questions and comments from looking at your submission:

- If this is a package for doing NARMAX, should NARMAX appear in the name of the package?
- How does this compare to something like [PySindy](https://joss.theoj.org/papers/10.21105/joss.02104)?
- Since this seems to be a way of inferring a model from data, how is it related to e.g. Scikit-Learn?

About the paper:
- You use `F` sometimes and `ℱ` (script F) other times.
- Perhaps there's a way of abbreviating equation (2)?
- It is not clear what type of objects the Θ's are.
- It is not clear where the degree ℓ comes into the equation.
- It's not clear why it's relevant that polynomial functions are smooth and the Weierstrass theorem.
- You don't say what SISO and MISO, AIC etc. are.  Maybe that can be left to the documentation, but you should say that.
- provid -> provide
- It is not clear what the table and figure are telling me. I.e. what is it exactly that your package outputs and how can I use that?

Basically I need your paper to explain what the package provides for a non-expert in the field (me).

Thanks!
