### 2020-05-28 203c3b6
update information criteria method
