### 2020-05-28 aaeedab
corrections and suggestions by the editor
