+++ paper/paper.bib
@@ -131,3 +131,27 @@
   pages={633--639},
   year={1885}
 }
+
+@article{KST2020,
+  title={A Tree Adjoining Grammar Representation for Models Of Stochastic Dynamical Systems},
+  author={Khandelwal, Dhruv and Schoukens, Maarten and T{\'o}th, Roland},
+  journal={arXiv preprint arXiv:2001.05320},
+  year={2020}
+}
+
+@inproceedings{LJAM2017,
+  title={Identificação de um Motor/Gerador CC por meio de Modelos Polinomiais Autorregressivos e Redes Neurais Artificais},
+  author={Lacerda Junior, W. R, L. and Almeida, V. M. and Martins, S. A. M.},
+  booktitle     = {XIII Simp{\'{o}}sio Brasileiro de Automa{\c{c}}{\~{a}}o Inteligente},
+  year          = {2017},
+  pages         = {1--6},
+  address       = {Porto Alegre}
+}
+
+@book{HK1999,
+  title={Nonlinear system identification. 2. Nonlinear system structure identification},
+  author={Haber, Robert and Keviczky, L{\'a}szl{\'o}},
+  volume={7},
+  year={1999},
+  publisher={Springer Science \& Business Media}
+}
