### 2020-05-28 #2191 @wilsonrljr
Hi @dpsanders ! Thank you for your comments and efforts towards improving our manuscript. In the following I will address all your questions and suggestions.

- If this is a package for doing NARMAX, should NARMAX appear in the name of the package?
In fact, adding NARMAX to the title would be a good way to make it clearer what the package is about.

- How does this compare to something like PySindy?
I take a look at PySindy and it is a great package for System Identification. However, although SysIdentPy and PySindy belong to the same field, they are very different model class. PySindy deals with autonomous systems like Lorenz System and other chaotic systems based on numerical integration and differentiation methods for optimization.

On the other hand, SysIdentPy is based on nonlinear autoregressive models with multiple inputs and outputs. To build the models we rely on algorithms like Forward Orthogonal Least Squares Algorithms and Error Reduction Ratio. The core of NARMAX models is the model structure selection do select the correct terms using methods such as the aforementioned algorithms. At this point we have polynomial NARMAX (the most used basis function for NARMAX), but we already implementing Neural NARMAX, for example. In short, SysIdenPy and PySindy provide very different ways to build dynamical models based on different goals, assumptions and methods.

 - Since this seems to be a way of inferring a model from data, how is it related to e.g. Scikit-Learn?
SysIdenPy provides built-in and native methods for building nonlinear autoregressive models using traditional algorithms in System Identification field. For modeling dynamical data using parametric models, some specific methods are required. Scikit-Learn have useful tools for building models from static data since but they not support dynamical data modeling. NARMAX models are the most used model class for building dynamical models. SysIdentPy offer native solutions and algorithms regarding NARMAX models including the model structure selection algorithms, the model order selection algorithms, several methods for parameter estimation and many other features.

 Actually, SysIdentPy have more in common with StatsModels package in terms of models class. But statsmodels only support **linear** autregressive models (ARMAX).  Nonlinear models requires robust model structure selection and parameter estimation algorithms, whice are provided in SysIdentPy.

In addition, SysIdentPy package is designed to be accessible to scikit-learn users while advanced users can customize it in many ways for different analysis.

- You use F sometimes and ℱ (script F) other times.
Thank you. The correction has been made.

- Perhaps there's a way of abbreviating equation (2)?
We changed it to a  shortened version of the equation.

- It is not clear what type of objects the Θ's are.
- It is not clear where the degree ℓ comes into the equation.
We added an example showing what are the $\Theta$ and $\ell$. Actually, we removed $\ell$ from equation 1 to use $\mathcal{F}$ to represent any function and we explain the polynomial NARMAX with more detail from equation 2 to make it more easy to understand.

- It's not clear why it's relevant that polynomial functions are smooth and the Weierstrass theorem.
The relevance relies on the fact that polynomials functions can be used to represent several nonlinear systems. As we mentioned in the paper, NARMAX models can be represented in a polynomial basis, neural basis, wavelet and so on. Although we are already implementing extensions like neural NARMAX, we like to highlight those polynomial NARMAX features because System Identification is not just about prediction, but can be also used for analysis of stability and control of nonlinear systems both in discrete and continuous time. In this respect, those kind of features make polynomial NARMAX powerful alternatives to for example neural, rbf NARMAX because you can easily access the model structure, take the derivatives, obtain an inverse model analytically and so on.

- You don't say what SISO and MISO, AIC etc. are. Maybe that can be left to the documentation, but you should say that.
Thank you for point out. We added more information in the respective paragraph as well some references for make it more clear. Also, we advise the user to read the package documentation for a more in depth details of those methods.

-  provid -> provide
The correction has been made.

- It is not clear what the table and figure are telling me. I.e. what is it exactly that your package outputs and how can I use that?
Thank you for this comment. We rewrote the paragraph explaining in more detail the procedure presented.

Note: Due to the mentioned changes, we have also have adapted some small parts here and there in the paper. If you need me to specify all the small changes, let me know.

If you have more suggestions, just tell me!
