### 2020-06-23 #2191 @dpsanders
👋 Hi @Shibabrat and @dawbarton. Would either of you be willing and able to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Thanks!
