### 2020-06-23 #2191 @dawbarton
Yes, I'd be willing to. I've worked a little with NARMAX models in the past.
