### 2020-07-10 #2384 @dawbarton
Overall, looks good. It installs easily and usage is as described. Documentation is good overall.

A few detailed comments below.

* Paper - some comments scribbled onto the PDF [10.21105.joss.02384 1.pdf](https://github.com/openjournals/joss-reviews/files/4904481/10.21105.joss.02384.1.pdf); it's a bit sloppy in places and looks like it really needed proof reading again.
* Documentation
    * `introduction_to_narmax.html`: $F^l$ in the "So, what is a NARMAX model?" section appears to be slightly misleading. The notation seems to indicate a polynomial function (particularly with the comment about setting $l=1$ recovering the ARMAX model) whereas you describe it as a general nonlinear function. In the section that follows, you refer to "unknown mapping $f[⋅]$"; I presume this was the $F^l$ in the previous section.
    * `user_guide.html`: the formatting of the text goes awry about halfway down (all the text appears to turn into headings). The text is fine, though you might want to consider how it is structured. I would recommend putting the function you are describing first and then explain it, rather than give a (rather specific) explanation and then the function. For example, the `get_siso_data` function - you start by giving an equation then say that `color_noise` equal to `True` means that you get a certain noise relationship. In the context of the equation (alone) that statement is confusing - what has `color_noise` got to do with the equation? It doesn't have any variables called `color_noise` in it. It's not until the end of the section that you see that the text was actually talking about a particular *function* (that is, `get_siso_data`) to generate sample data.
    * I'm not sure what the Jupyter notebooks section of the docs is supposed to do - it appears to be largely the same as the user guide.
* I can't see any automated tests in the repo - have I missed them or do they not exist? I think JOSS requires them?
