### 2020-07-17 #2384 @wilsonrljr
Dear @dawbarton , we appreciate the time and effort that you dedicated to providing feedback on our manuscript.

- Paper - some comments scribbled onto the PDF 10.21105.joss.02384 1.pdf; it's a bit sloppy in places and looks like it really needed proof reading again.
Regarding the manuscript, we are incorporating the suggestions made by you.  The comments will result in valuable improvements to our paper.

- introduction_to_narmax.html: $F^l$ in the "So, what is a NARMAX model?" section appears to be slightly misleading. The notation seems to indicate a polynomial function (particularly with the comment about setting $l=1$ recovering the ARMAX model) whereas you describe it as a general nonlinear function. In the section that follows, you refer to "unknown mapping $f[⋅]$"; I presume this was the $F^l$ in the previous section.
Thank you for pointing this out. The $F$ in NARMAX indicates a general nonlinear function (Neural, Fuzzy, and many other). However, because we define it as a polynomial function, we have to keep the definition consistent. I will upload the changes soon.

- user_guide.html:  the formatting of the text goes awry about halfway down (all the text appears to turn into headings). The text is fine, though you might want to consider how it is structured.
We agree with you. The documentation is in a very early stage. Actually, the website version will be replaced very soon. Many issues with structure and formatting is due the current way it is done. We are converting the docs to a properly website.

In this case, I have a question: Would these changes concerning the documentation in the website be mandatory in this revision process?

- I'm not sure what the Jupyter notebooks section of the docs is supposed to do - it appears to be largely the same as the user guide.
 Each notebook presents a different exemple of how particular parameters works and some tips. We made this section for the users who have no prior knowledge of the library or NARMAX models. There are particular usage of some parameters we didn't show in user guide.

- I can't see any automated tests in the repo - have I missed them or do they not exist? I think JOSS requires them?
We have tests that coverage approximately 87% of the our functions. These tests are placed on "tests" folder inside each class path.
