+++ paper/paper.bib
@@ -17,7 +17,7 @@
 @InProceedings{BL1981,
   author    = {Billings, S. A. and Leontaritis, I. J.},
   title     = {{Identification of Nonlinear Systems Using Parameter Estimation Techniques}},
-  booktitle = {Proceedings of the IEEE Conference on Control and its Application},
+  booktitle = {Proceedings of the {IEEE} Conference on Control and its Application},
   year      = {1981},
   pages     = {183--187}
 }
@@ -154,4 +154,14 @@
   volume={7},
   year={1999},
   publisher={Springer Science \& Business Media}
+}
+
+@article{ayala2020r,
+  title={An R library for nonlinear black-box system identification},
+  author={Ayala, Helon Vicente Hultmann and Gritti, Marcos Cesar and dos Santos Coelho, Leandro},
+  journal={SoftwareX},
+  volume={11},
+  pages={100495},
+  year={2020},
+  publisher={Elsevier}
 }
