### 2020-08-14 108bd9e
small update in a documentation example
