### 2020-08-14 c6618a2
Create CODE_OF_CONDUCT.md
