### 2020-08-14 #2384 @dpsanders
@wilsonrljr: I was wondering if http://www.nonlinearbenchmark.org/ is relevant?
