### 2020-08-14 77f3c09
add example of f_16 benchmark
