### 2020-08-14 #2384 @wilsonrljr
Thanks for pointing that out, dawbarton. I worked with some of those datasets on my master dissertation. Some of the authors collaborating with the website are reference on system identification using NARMAX models. It's really important to share our work with researchers on the field with examples that make sense for them.

I updated the master branch with the F-16 example. It's a simple showcase for now, without any comparison with other works.  https://wilsonrljr.github.io/sysidentpy/examples/f_16_benchmark.html
