### 2020-09-10 #2384 @Shibabrat
@dawbarton I've finished my review and satisfied with the current version of the package.

@wilsonrljr can you add a section on testing to the README page? I see the tests are available but it may not be obvious to a first time user. Otherwise, LGTM.
