+++ README.md
@@ -1,5 +1,3 @@
-.. -*- mode: rst -*-
-
 SysIdentPy
 ==========
 
