### 2020-09-11 f4d13d7
minor changes in readme.md
