### 2020-09-11 #2384 @wilsonrljr
Thanks, @Shibabrat!

Following your suggestion, I added the following note to the README on Development section:

"Note: we use the pytest package for testing. The test functions are located in tests subdirectories at each folder inside sysidentpy, which check the validity of the algorithms.

Run the pytest in the respective folder to perform all the tests of the corresponding sub-packages.

Currently, we have around 85% of code coverage."
