### 2020-09-12 1d0ed74
minor changes in README
