+++ README.md
@@ -1,105 +1,128 @@
-SysIdentPy
-==========
 
-sysidentpy is a Python module for System Identification using NARMAX models built on top of numpy and
-is distributed under the 3-Clause BSD license.
+# SysIdentPy
 
-The project was started in by Wilson R. L. Junior, Luan
-Pascoal C. Andrade and Samir A. M. Martins as a project for System Identification
-discipline. Samuel joined early in 2019 and since then have contributed.
+SysIdentPy is a Python module for System Identification using **NARMAX** models built on top of **numpy** and is distributed under the 3-Clause BSD license.
 
-Documentation
-=============
+The project was started in by Wilson R. L. Junior, Luan Pascoal C. Andrade and Samir A. M. Martins as a project for System Identification discipline. Samuel joined early in 2019 and since then have contributed.
+
+# Documentation
 
 - Website: https://wilsonrljr.github.io/sysidentpy/
 
-Examples
---------
+# Examples
+
 The examples directory has several Jupyter notebooks presenting basic tutorials of how to use the package and some specific applications of sysidentpy. Try it out!
 
-Requirements
-------------
+# Requirements
 
-sysidentpy requires:
+SysIdentPy requires:
 
 - Python (>= 3.6)
 - NumPy (>= 1.5.0) for all numerical algorithms
 - Matplotlib >= 1.5.2 for static plotiing and visualizations
 
-| Platform  | Status |
+| Platform | Status |
 | --------- | -----:|
-| Linux     | ok    |
-| Windows   | ok    |
-| macOS     | ok    |
-
+| Linux | ok |
+| Windows | ok |
+| macOS | ok |
 
 **SysIdentPy do not to support Python 2.7.**
 
 A few examples require pandas >= 0.18.0. However, it is not required to use sysidentpy.
 
-Installation
-============
+# Installation
 
-The easiest way to get sysidentpy running is to install it using ``pip``   ::
-
-    pip install sysidentpy
+The easiest way to get sysidentpy running is to install it using ``pip``
+~~~~~~~~~~~~~~~~
+pip install sysidentpy
+~~~~~~~~~~~~~~~~
 
 We will made it available at conda repository as soon as possible.
 
-Changelog
-=========
+# Changelog
 
-See the `changelog <https://github.com/wilsonrljr/sysidentpy.whats_is_new.md>`__
-for a history of notable changes to sysidentpy.
+See the [changelog]( <https://github.com/wilsonrljr/sysidentpy.whats_is_new.md>) for a history of notable changes to SysIdentPy.
 
-Development
-===========
+# Development
 
 We welcome new contributors of all experience levels. The sysidentpy community goals are to be helpful, welcoming, and effective.
 
-Note: we use the pytest package for testing. The test functions are located in tests subdirectories at each folder inside sysidentpy, which check the validity of the algorithms.
+*Note*: we use the `pytest` package for testing. The test functions are located in tests sub directories at each folder inside **SysIdentPy**, which check the validity of the algorithms.
 
-Run the pytest in the respective folder to perform all the tests of the corresponding sub-packages.
+Run the `pytest` in the respective folder to perform all the tests of the corresponding sub-packages.
 
-Currently, we have around 85% of code coverage.
+Currently, we have around 81% of code coverage.
 
-Important links
-===============
+You can install pytest using
+~~~~~~~~~~~~~~~~~~~~~~
+pip install -U pytest
+~~~~~~~~~~~~~~~~~~~~~~
+
+### Example of how to run the tests:
+
[…]
