### 2020-09-12 0390c3b
minor fixes on readme
