### 2020-09-12 #2384 @wilsonrljr
I added more information concerning the tests, including an example of how to run it. If you have any more suggestions, just let me know!
