### 2020-09-12 64031e7
Minor edits to paper
