### 2020-09-12 #2384 @dpsanders
Once the final details have been fixed, I think we're ready to proceed to publication!

Could you please make a tagged release and archive the software (e.g. on Zenodo or Figshare), and report the version number and archive DOI here. Thanks!
