### 2020-09-12 #2384 @wilsonrljr
I made a new release (release name -> SysIdentPy v0.1.2, version v0.1.2).

Zenodo registered two DOIs:
- DOI for v0.1.2 -> DOI: 10.5281/zenodo.4026516;
- DOI for all versions -> DOI 10.5281/zenodo.4026515
