### 2020-09-12 efc5222
update to user_guide doc
