+++ docs/_sources/user_guide.rst.txt
@@ -24,11 +24,10 @@ Generating 1 input 1 output sample data
 ---------------------------------------
 
 The data is generated by simulating the following model:
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 :math:`y_k = 0.2y_{k-1} + 0.1y_{k-1}x_{k-1} + 0.9x_{k-1} + e_{k}`
 
-If *colored_noise* is set to True:
+If ``colored_noise`` is set to True:
 
 :math:`e_{k} = 0.8\nu_{k-1} + \nu_{k}`
 
@@ -46,26 +45,24 @@ noise and selecting 90% of the data to train the model.
                                                        sigma=0.001,
                                                        train_percentage=90)
 
-To obtain a NARMAX model we have to choose some values, *e.g*, the nonlinearity degree (*non_degree*), the maximum lag for the inputs and output (*xlag* and *ylag*).
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+To obtain a NARMAX model we have to choose some values, e.g, the nonlinearity degree (``non_degree``), the maximum lag for the inputs and output (``xlag`` and ``ylag``).
 
 In addition, you can select the information criteria to be used with the Error Reduction Ratio to select the model order and the method to estimate the model paramaters:
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 -  Information Criteria: aic, bic, lilc, fpe
 -  Parameter Estimation: least_squares, total_least_squares,
    recursive_least_squares, least_mean_squres and many other (see the
    docs)
 
-The *n_terms* values is optional. It refer to the number of terms to
+The ``n_terms`` values is optional. It refer to the number of terms to
 inclued in the final model. You can set this value based on the
 information criteria (see below) or based on priori information about
-the model struture. The default value is *n_terms=None*, so the
+the model struture. The default value is ``n_terms=None``, so the
 algorithm will choose the minimum value reached by the information
 criteria.
 
-To use information criteria you have to set *order_selection=True*. You
-can also select *n_info_values* (default = 15).
+To use information criteria you have to set ``order_selection=True``. You
+can also select ``n_info_values`` (default = 15).
 
 .. code:: ipython3
 
@@ -78,8 +75,7 @@ can also select *n_info_values* (default = 15).
                              estimator='least_squares',
                              )
 
-The *fit* method executes the Error Reduction Ratio algorithm using Househoulder reflection to select the model structure.
[…]
