### 2020-09-27 #2384 @dpsanders
Thanks! and sorry for the delay here.

I'm going to proceed towards publication!
