+++ paper/paper.bib
@@ -2,7 +2,8 @@
   title={System identification: theory for the user},
   author={Ljung, Lennart},
   year={1987},
-  publisher={Prentice-hall}
+  publisher={Prentice-hall},
+  doi={10.1016/0005-1098(89)90019-8}
 }
 
 @Book{Bil2013,
@@ -41,7 +42,8 @@
   number={3},
   pages={1013--1032},
   year={1989},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/00207178908559683}
 }
 
 @article{MA2016,
@@ -51,7 +53,8 @@
   volume={75},
   pages={607--617},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ymssp.2015.12.031}
 }
 
 @book{Agu2004,
@@ -68,7 +71,8 @@
   volume={7},
   pages={98228--98237},
   year={2019},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/access.2019.2926057}
 }
 
 @article{FWHM2003,
@@ -79,7 +83,8 @@
   number={8},
   pages={611--627},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/s0307-904x(03)00071-4}
 }
 
 @article{GGBW2016,
@@ -89,7 +94,8 @@
   volume={173},
   pages={715--723},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.neucom.2015.08.022}
 }
 
 @article{KGHK2003,
@@ -100,7 +106,8 @@
   number={1},
   pages={70--81},
   year={2003},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tbme.2002.803507}
 }
 
 @incollection{BBWL2018,
@@ -109,7 +116,8 @@
   booktitle={Machine Learning Techniques for Space Weather},
   pages={203--236},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/b978-0-12-811788-0.00008-1}
 }
 
 @article{CER2001,
@@ -136,7 +144,8 @@
   title={A Tree Adjoining Grammar Representation for Models Of Stochastic Dynamical Systems},
   author={Khandelwal, Dhruv and Schoukens, Maarten and T{\'o}th, Roland},
   journal={arXiv preprint arXiv:2001.05320},
-  year={2020}
+  year={2020},
+  doi={10.1016/j.automatica.2020.109099}
 }
 
 @inproceedings{LJAM2017,
