### 2020-10-01 #2384 @Kevin-Mattheus-Moerman
@wilsonrljr I've reviewed your paper and ask you to make these last minor changes:

- [ ] Please add a country to your affiliation

- [ ] The figure axis font for figure 1 is very small for the bottom figure. I recommend you increase the font and update the picture.

- [ ] Please edit the ZENODO archive meta data to match that of the paper, i.e. the title and author list has to match that of the paper. (@dpsanders for future reference please check this before proceeding to recommend acceptance, thanks. )

- [ ] Your paper is about to be processed for acceptance. Please thoroughly read through it once more yourself. .

Let me know when you've made these changes. Thanks.

Kevin Moerman
_AEiC JOSS_
