### 2020-10-02 6913ff3
figure 1 on paper was updated; affiliation included
