### 2020-10-02 #2384 @wilsonrljr
@Kevin-Mattheus-Moerman Thank you for reviewing the paper.
- I added a country affiliation;
- I've increased the font size of the figure 1. Now all fonts have the same size on labels;
- I updated the title on Zenodo metadata to match the paper title. The author list was already the same of the paper, but I've included the country affiliation on Zenodo too.

Let me know if you need anything else.
