### 2020-10-02 #2384 @wilsonrljr
Many thanks @dpsanders, @dawbarton, @Shibabrat and @Kevin-Mattheus-Moerman!  All of you have contributed to the effort to obtain the this version of the paper.
